+++ include/roq/order_update.hpp
@@ -46,6 +46,7 @@ struct ROQ_PUBLIC OrderUpdate final {
   std::chrono::nanoseconds update_time_utc = {};      //!< Updated timestamp (UTC)
   std::string_view external_account;                  //!< External account name
   std::string_view external_order_id;                 //!< External order identifier
+  std::string_view client_order_id;                   //!< Client order identifier
   OrderStatus status = {};                            //!< Order status
   double quantity = NaN;                              //!< Quantity (total, indicative)
   double price = NaN;                                 //!< Price
@@ -101,6 +102,7 @@ struct fmt::formatter<roq::OrderUpdate> {
         R"(update_time_utc={}, )"
         R"(external_account="{}", )"
         R"(external_order_id="{}", )"
+        R"(client_order_id="{}", )"
         R"(status={}, )"
         R"(quantity={}, )"
         R"(price={}, )"
@@ -134,6 +136,7 @@ struct fmt::formatter<roq::OrderUpdate> {
         value.update_time_utc,
         value.external_account,
         value.external_order_id,
+        value.client_order_id,
         value.status,
         value.quantity,
         value.price,